refactor: extract persistence logic around tags
