+++ lib/services/link_service.dart
@@ -1,7 +1,8 @@
 import 'package:flutter/services.dart';
 import 'package:lotti/database/database.dart';
+import 'package:lotti/features/tags/repository/tags_repository.dart';
 import 'package:lotti/get_it.dart';
-import 'package:lotti/logic/persistence_logic.dart';
+import 'package:lotti/logic/persistence/persistence_logic.dart';
 import 'package:lotti/services/tags_service.dart';
 
 class LinkService {
@@ -25,7 +26,7 @@ class LinkService {
       final linkedTagIds = linkedFrom?.meta.tagIds;
       final storyTags = _tagsService.getFilteredStoryTagIds(linkedTagIds);
 
-      await _persistenceLogic.addTags(
+      await TagsRepository.addTags(
         journalEntityId: _linkToId!,
         addedTagIds: storyTags,
       );
